# Conflicts:
#	src/main/resources/application.yaml

+++ src/main/kotlin/com/brifo/server/diary/dto/response/GetDiaryDetailResponse.kt
@@ -2,6 +2,7 @@ package com.brifo.server.diary.dto.response
 
 import com.brifo.server.agent.entity.AgentType
 import com.brifo.server.briefing.entity.BriefingDirection
+import java.math.BigDecimal
 import java.util.UUID
 
 data class GetDiaryDetailResponse(
@@ -15,7 +16,7 @@ data class GetDiaryDetailResponse(
     data class Stock(
         val stockId: UUID,
         val name: String,
-        val changeRate: Int,
+        val changeRate: BigDecimal,
     )
 
     data class Agent(
